backend/requirements.txt
fastapi
uvicorn
assemblyai
together
python-dotenv
